pyproject.toml
glm-ocr-gui = { path = ".", editable = true }

[project]
name = "glm-ocr-gui"
version = "0.1.0"
description = "GLM-OCR Drag and Drop GUI"
requires-python = ">=3.10"
dependencies = [
    "PyQt5>=5.15.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
]

[tool.uv]
dev-dependencies = []